feat(Core/Entities): Make rage generation configurable
Pulls the retail 3.3.5 rage formulas out of Unit.cpp into
Acore::Rage in Formulas.h so they can be unit tested, and adds three
config options: Rage.Normalized, Rage.NormalizedMultiplier and
Rage.FromDamageTaken.

The conversion denominator is quadratic only up to level 70; past that
it is a linear WotLK-era guess rather than a datamined constant, which
is what makes level 80 the awkward case for any rage retuning. Solo
play needs a lever here because a Protection warrior with no other
players feeding it damage generates rage far below what the tree
assumes.

FormulasTest.cpp covers both formulas, including the cap that stops a
slow weapon landing a trivial hit from paying out.

+++ src/server/game/Entities/Unit/Unit.h
@@ -1143,7 +1143,7 @@ public:
 
     int32 ModifyPower(Powers power, int32 val, bool withPowerUpdate = true);
 
-    void RewardRage(uint32 damage, uint32 weaponSpeedHitFactor, bool attacker);
+    void RewardRage(uint32 damage, float weaponSpeedHitFactor, bool attacker);
 
     [[nodiscard]] uint32 GetCreateMana() const { return GetUInt32Value(UNIT_FIELD_BASE_MANA); }
     void SetCreateMana(uint32 val) { SetUInt32Value(UNIT_FIELD_BASE_MANA, val); }
